Refactor GUI related classes; clean up FlySpeed enum logic
