+++ src/main/java/tc/oc/pgm/observers/tools/GamemodeTool.java
@@ -12,10 +12,10 @@ import tc.oc.component.render.ComponentRenderers;
 import tc.oc.component.types.PersonalizedText;
 import tc.oc.component.types.PersonalizedTranslatable;
 import tc.oc.pgm.api.player.MatchPlayer;
-import tc.oc.pgm.gui.InventoryGUI;
-import tc.oc.pgm.observers.ObserverTool;
+import tc.oc.pgm.gui.InventoryMenu;
+import tc.oc.pgm.gui.InventoryMenuItem;
 
-public class GamemodeTool implements ObserverTool {
+public class GamemodeTool implements InventoryMenuItem {
 
   @Override
   public Component getName() {
@@ -45,7 +45,7 @@ public class GamemodeTool implements ObserverTool {
   }
 
   @Override
-  public void onInventoryClick(InventoryGUI menu, MatchPlayer player) {
+  public void onInventoryClick(InventoryMenu menu, MatchPlayer player) {
     toggleObserverGameMode(player);
     menu.refreshWindow(player);
   }
